Move ProjectDBManager class from manager.py to project_db_manager.py
